20191028 Update Account Manage Service

+++ src/service/AccountDepositWithdraw.java
@@ -1,8 +1,10 @@
 package service;
 
+import java.math.BigDecimal;
+
 public interface AccountDepositWithdraw {
 
-    public int Deposit(String id, double amount);
+    public int Deposit(String id, BigDecimal amount);
 
-    public int Withdraw(String id, double amount);
+    public int Withdraw(String id, BigDecimal amount);
 }
